H2/requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
python-dotenv==1.0.1
Werkzeug==3.0.4
zipp==3.20.2
# python-dotenv
# ipykernel
# torch
# pillow
# transformers
# openai
langchain-openai
langchain_community
langchain_chroma
pandas
langchain-huggingface

